Fix Missing USB Traces (#74)
This fixes a massive issue I created for myself last year. I made the
assumption that the matching D+/- pins on the USB-C connecter were
electrically tied in the connector, and therefore I did not have to tie
them together myself. This was incorrect and resulted in having to
insert two small jumpers in every one of the 125 production boards which
was pure pain.